router made private
